requirements.txt
# requirements.txt
# =================
# Python dependencies for the ANE Cache Research pipeline.
# Install: pip install -r requirements.txt
#
# Note: coremltools requires macOS with Xcode installed for model conversion.
# All analysis scripts will run on any OS (Linux, Windows, macOS).

# ── Model conversion ──────────────────────────────────────────────────────────
coremltools>=7.0
torch>=2.0.0
torchvision>=0.15.0

# ── Data processing ───────────────────────────────────────────────────────────
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
pyyaml>=6.0

# ── Statistical analysis ──────────────────────────────────────────────────────
pingouin>=0.5.4        # ANOVA + effect sizes
statsmodels>=0.14.0    # OLS, ANOVA fallback

# ── Visualisation ─────────────────────────────────────────────────────────────
matplotlib>=3.7.0
seaborn>=0.12.0        # optional — used in notebooks

# ── Notebooks ─────────────────────────────────────────────────────────────────
jupyter>=1.0.0
ipykernel>=6.0.0

# ── Dev / utilities ───────────────────────────────────────────────────────────
tqdm>=4.65.0           # progress bars
rich>=13.0.0           # pretty terminal output
